moved countries stuff to helpers
