+++ client/src/trip/components/TripForm.js
@@ -2,10 +2,10 @@ import { useState } from "react";
 import { v4 as uuidv4 } from "uuid";
 import initialTrip from "../reducers/initialTrip";
 import { TextField } from "@mui/material";
-import SelectCountry from "./TripFormCountry";
 import FormGroup from "@mui/material/FormGroup";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
+import SelectCountry from "../../helpers/select-country/select-country.js"
 
 const TripForm = ({ handleAddTrip }) => {
   const [data, setData] = useState({ ...initialTrip, IsPublic: true });
